create test_environment
